# Migrate github.com/json-iterator/go to sigs.k8s.io/json