--- a/go.mod
+++ b/go.mod
@@ -3,10 +3,9 @@ module sigs.k8s.io/structured-merge-diff/v4
 require gopkg.in/yaml.v2 v2.2.8
 
 require (
-	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/google/go-cmp v0.5.9
 	github.com/google/gofuzz v1.0.0
-	github.com/json-iterator/go v1.1.12
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd
 )
 
 go 1.13