fixes https://github.com/kubernetes-sigs/structured-merge-diff/issues/202

Replaces "github.com/json-iterator/go" with "sigs.k8s.io/json" and "encoding/json".

TODO: compare benchmarks before and after <- help would be appreciated here.