@apelisse 2024-04-11
Yeah, that's ambitious. There's a reason we had written the code the way we did, but let's see what the benchmarks say.

You can do ... IIRC
```
go test -bench= ./fieldpath/
```
to run the benchmark. Do that against master and the tip of your branch, save the output in files, run `benchstat master.txt pr.txt`, see what comes out (you might need to do add a `-count=10` to your `go test`.