@jpbetz 2024-04-11
Benchmarks will be essential here. This code was hyper-optimized because it was on a hot path.