@jpbetz 2024-04-11
/hold
Thanks @apelisse!  Holding based on those numbers to make sure we don't cause a performance regression.